+++ src/traits.rs
@@ -2,7 +2,7 @@ use std::result::Result as StdResult;
 
 use reqwest::Error as ReqwestError;
 use reqwest::Response;
-use serde::{Serialize, Deserialize};
+use serde::{de, Serialize};
 
 use {Error, Result};
 use errors::RestError;
@@ -18,10 +18,10 @@ pub trait ApiResponse : Sized {
     fn validate_status(self) -> Result<Response>;
     
     /// Attempts to parse the response as JSON into `T`.
-    fn deserialize<T: Deserialize>(self) -> Result<T>;
+    fn deserialize<T: de::DeserializeOwned>(self) -> Result<T>;
     
     /// Validates the status code and attempts to deserialize into `T` in one step.
-    fn validate_and_read<T: Deserialize>(self) -> Result<T> {
+    fn validate_and_read<T: de::DeserializeOwned>(self) -> Result<T> {
         self.validate_status()?.deserialize()
     }
 }
@@ -35,7 +35,7 @@ impl ApiResponse for Response {
         }
     }
     
-    fn deserialize<T: Deserialize>(self) -> Result<T> {
+    fn deserialize<T: de::DeserializeOwned>(self) -> Result<T> {
         self.validate_status()?.json().map_err(Error::from)
     }
 }
@@ -45,7 +45,7 @@ impl ApiResponse for StdResult<Response, ReqwestError> {
         self.map_err(Error::from).and_then(ApiResponse::validate_status)
     }
     
-    fn deserialize<T: Deserialize>(self) -> Result<T> {
+    fn deserialize<T: de::DeserializeOwned>(self) -> Result<T> {
         self.validate_status().and_then(ApiResponse::deserialize)
     }
 }